AI_Employee_Vault/gold/requirements.txt
# Gold Tier Python Dependencies

# MCP (Model Context Protocol)
mcp[cli]>=0.9.0

# Core
python-dotenv>=1.0.0
pydantic>=2.5.0
pyyaml>=6.0.1

# Async & Concurrency
asyncio>=3.4.3
aiohttp>=3.9.0

# Social Media APIs
requests>=2.31.0
facebook-sdk>=3.1.0
python-twitter>=3.5
instagrapi>=2.0.0

# Date/Time
python-dateutil>=2.8.2
pytz>=2023.3

# Logging & Monitoring
structlog>=23.2.0
colorlog>=6.8.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Development
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.1
isort>=5.13.0

# Process Management
psutil>=5.9.6

# Utilities
tenacity>=8.2.3  # Retry logic
jsonschema>=4.20.0  # Schema validation
